Fix Sphinx documentation issues.
