+++ python/BioSimSpace/Metadynamics/CollectiveVariable/_distance.py
@@ -372,8 +372,8 @@ class Distance(_CollectiveVariable):
         """Get the weights to be used when computing the center of the first
            atom group.
 
-           Returnas
-           --------
+           Returns
+           -------
 
            weights0 : [float]
                A list of weights to be used when computing the center of the
@@ -432,8 +432,8 @@ class Distance(_CollectiveVariable):
         """Get the weights to be used when computing the center of the second
            atom group.
 
-           Returnas
-           --------
+           Returns
+           -------
 
            weights1 : [float]
                A list of weights to be used when computing the center of the
